+++ lib/rley.rb
@@ -1,0 +1,9 @@
+# File: rley.rb
+# This file acts as a jumping-off point for loading dependencies expected
+# for a Rley client.
+
+require_relative './rley/constants'
+require_relative './rley/syntax/grammar_builder'
+require_relative './rley/parser/earley_parser'
+
+# End of file